--- frontend-osori/src/components/layout/Bottombar.vue
@@ -1,1 +1,0 @@
-<!--Bottombar.vue 더 컴포넌트 재사용하기 편하게 디렉토리 수정할 것-->
