+++ frontend-osori/src/router/index.js
@@ -1,34 +1,40 @@
 /* eslint-disable */
 import { createRouter, createWebHashHistory } from 'vue-router'
-import HomeView from '../views/HomeView.vue'
-import Profile from '@/views/Profile.vue'
 import AuthLayout from '@/components/layout/AuthLayout.vue' // 중첩 라우팅의 부모 레이아웃
-import VerifyEmail from '@/views/login/VerifyEmail.vue'
-import SignIn from '@/views/login/SignIn.vue'
-import SignUpForm from '@/views/login/SignUpForm.vue'
-import PortfolioCreate from '@/views/portfolio/PortfolioCreate.vue'
-import PortfolioByTag from '@/views/portfolio/PortfolioByTag.vue'
-import PortfolioTest from '@/views/portfolio/PortfolioTest.vue'
 import PortfolioLayout from '@/components/layout/PortfolioLayout.vue' // 중첩 라우팅의 부모 레이아웃
-import TaliwindResopon from '@/views/TaliwindResopon.vue'
-import SignUpForm2 from '@/views/login/SignUpForm2.vue'
-import NotFound from '@/views/NotFound.vue'
-import WeeklyCalenderTest from '@/views/WeeklyCalenderTest.vue'
-import realsignUp from '@/views/login/realsignup.vue'
-import realLogin from '@/views/login/realLogin.vue'
-//import NotFoundPage from '@/views/NotFoundPage.vue' // 404 페이지
+
+// 자동 임포트 함수 (src/views 내의 모든 .vue 파일을 임포트)
+function importAllViews() {
+  const viewFiles = require.context('@/views', true, /\.vue$/)
+  const views = {}
+
+  // 각 파일을 순회하면서 임포트
+  viewFiles.keys().forEach((filePath) => {
+    const viewName = filePath
+      .split('/')
+      .pop() // 파일 이름만 추출
+      .replace('.vue', '') // .vue 확장자 제거
+
+    // 컴포넌트 이름과 해당 컴포넌트를 맵핑
+    views[viewName] = viewFiles(filePath).default
+  })
+
+  return views
+}
+
+// 자동으로 임포트된 모든 Vue 컴포넌트 객체
+const importedViews = importAllViews()
 
 const routes = [
   {
     path: '/',
     name: 'home',
-    component: HomeView
+    component: importedViews['HomeView'] // 자동 임포트 적용
   },
   {
     path: '/about',
     name: 'about',
-    component: () =>
-      import(/* webpackChunkName: "about" */ '../views/AboutView.vue')
+    component: importedViews['AboutView'] // 자동 임포트 적용
   },
   {
     path: '/auth',
@@ -37,79 +43,77 @@ const routes = [
       {
         path: 'signin',
         name: 'signin',
-        component: SignIn
+        component: importedViews['SignIn'] // 자동 임포트 적용
       },
       {
         path: 'signupform',
         name: 'signupform',
-        component: SignUpForm
+        component: importedViews['SignUpForm'] // 자동 임포트 적용
       },
       {
         path: 'verifyemail',
         name: 'verifyemail',
-        component: VerifyEmail
+        component: importedViews['VerifyEmail'] // 자동 임포트 적용
       },
       {
-        path: 'signupform2', // /portfolios/tag
+        path: 'signupform2',
         name: 'SignUpForm2',
-        component: SignUpForm2
+        component: importedViews['SignUpForm2'] // 자동 임포트 적용
       },
       {
-        path: 'realsignup', // /portfolios/tag
+        path: 'realsignup',
         name: 'realsignUp',
-        component: realsignUp
+        component: importedViews['realsignup'] // 자동 임포트 적용
       },
       {
-        path: 'reallogin', // /portfolios/tag
+        path: 'reallogin',
         name: 'realLogin',
-        component: realLogin
+        component: importedViews['realLogin'] // 자동 임포트 적용
       }
     ]
   },
   {
     path: '/profile',
     name: 'profile',
-    component: Profile,
+    component: importedViews['Profile'], // 자동 임포트 적용
     meta: { requiresAuth: true } // 인증이 필요한 페이지
   },
-  // 잘못된 경로 처리 (404 페이지)
   {
-    path: '/:catchAll(.*)', // 모든 경로를 잡아내는 방식
-    name: 'NotFound',
-    component: NotFound
-  },
-  {
-    path: '/portfolios', // 부모 라우트
+    path: '/portfolios',
     component: PortfolioLayout, // 공통 레이아웃 컴포넌트
     children: [
       {
-        path: 'new', // /portfolios/new
+        path: 'new',
         name: 'CreatePortfolio',
-        component: PortfolioCreate
+        component: importedViews['PortfolioCreate'] // 자동 임포트 적용
       },
       {
-        path: 'tag', // /portfolios/tag
+        path: 'tag',
         name: 'SearchByTag',
-        component: PortfolioByTag
+        component: importedViews['PortfolioByTag'] // 자동 임포트 적용
       },
       {
-        path: 'test', // /portfolios/tag
+        path: 'test',
         name: 'Test',
-        component: PortfolioTest
+        component: importedViews['PortfolioTest'] // 자동 임포트 적용
       }
     ]
   },
   {
     path: '/responsivetest',
     name: 'TaliwindResopon',
-    component: TaliwindResopon
+    component: importedViews['TaliwindResopon'] // 자동 임포트 적용
   },
   {
     path: '/weeklycalendertest',
     name: 'WeeklyCalenderTest',
-    component: WeeklyCalenderTest
+    component: importedViews['WeeklyCalenderTest'] // 자동 임포트 적용
+  },
+  {
+    path: '/:catchAll(.*)',
+    name: 'NotFound',
+    component: importedViews['NotFound'] // 자동 임포트 적용
   }
-  //{ path: '*', component: NotFoundPage } // 404 페이지 처리 (활성화 필요시)
 ]
 
 const router = createRouter({
@@ -117,23 +121,6 @@ const router = createRouter({
   routes
 })
 
-// 라우트 가드
-router.beforeEach((to, from, next) => {
-  if (to.matched.some((record) => record.meta.requiresAuth)) {
-    // 인증이 필요한 페이지를 확인
-    if (!isAuthenticated()) {
-      next({ name: 'signin' }) // 로그인되지 않은 경우 로그인 페이지로 리다이렉트
-    } else {
-      next() // 인증된 경우 계속 진행
-    }
-  } else {
-    next() // 인증이 필요 없는 페이지는 그대로 진행
-  }
-})
-
-function isAuthenticated() {
-  // 로컬 스토리지에서 'user' 항목으로 로그인 여부를 확인하는 함수
-  return !!localStorage.getItem('user') // 'user' 정보가 있으면 true 반환
-}
+//라우터 가드 설정(세션 기반 쿠기, 나중에 수정 필요)
 
 export default router
