+++ frontend-osori/src/store/index.js
@@ -1,10 +1,9 @@
 import { createStore } from 'vuex'
-import auth from './modules/auth.js' // auth 모듈 import
+import auth from '../views/test/auth.js' // auth 모듈 import
 import login from './modules/login.js'
 
 const store = createStore({
   modules: {
-    auth, // auth 모듈 등록
     login
   }
 })
