+++ frontend-osori/src/views/test/VerifyEmail.vue
